@@ -4,7 +4,7 @@ version: 0.0.2
 homepage: https://github.com/wskfjtheqian/chinese_calendar
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
